refactor: 상품 예상 금액 계산 로직 분리
